start putting api calls in helper functions instead to simplify app.js code
